Update rollsum.[ch] to use C99 features.
Change rollsum.h preprocessor macros to "static inline" functions. Use
stdint.h types uint_fast16_t and uint32_t where appropriate. Use
stddef.h type size_t where appropriate.

Update rollsum_test.c to remove uint32_t cast that is nolonger needed.

+++ src/rollsum.c
@@ -27,12 +27,12 @@
 #define DO8(buf,i)  DO4(buf,i); DO4(buf,i+4);
 #define DO16(buf)   DO8(buf,0); DO8(buf,8);
 
-void RollsumUpdate(Rollsum *sum,const unsigned char *buf,unsigned int len) {
+void RollsumUpdate(Rollsum *sum, const unsigned char *buf, size_t len) {
     /* ANSI C says no overflow for unsigned. zlib's adler32 goes to extra
     effort to avoid overflow for its mod prime, which we don't have.*/
-    unsigned int n = len;
-    unsigned long s1 = sum->s1;
-    unsigned long s2 = sum->s2;
+    size_t n = len;
+    uint_fast16_t s1 = sum->s1;
+    uint_fast16_t s2 = sum->s2;
 
     while (n >= 16) {
         DO16(buf);
